requirements.txt
Django==3.0.6
psycopg2==2.8.5
pygresql
django-sass-processor==0.8
django-webpack-loader==0.7.0
django-redis==4.11.0
django-fakeredis==0.1.1
https://github.com/darklow/django-suit/tarball/v2
requests==2.22.0
django-encrypted-secrets==0.9.8